wip split API
